[#1065] Today·Month 순수 뷰와 ViewModel 을 WidgetScenes 로 내린다
- `TodaySummaryView`·`SingleMonthView` 와 `TodayWidgetViewModel`·`MonthWidgetViewModel` 이 `WidgetScenes` 로 갔다. 확장에는 `entry:` 를 받는 엔트리 뷰와 `Widget` 선언·Provider 만 남는다
- `TodayWidgetViewModel.updated(events:)` 는 확장에 extension 으로 남겼다. `CalendarEvents` 가 확장 소유 타입이라 ViewModel 본체와 함께 내려갈 수 없다
- `WidgetAppearanceSettings.Background.colorSet` 을 `WidgetBackgroundStyle` 옆으로 옮겼다. 순수 뷰 6곳이 쓰는 배경 해석이라 C2 가 `WidgetScenes` 로 지정한 대상이고, extension 형태를 유지해 호출 코드는 안 바뀐다
- `WidgetBackgroundColorSetTests` — system 추종·custom 우선·hex 파싱 실패 폴백 세 분기를 덮는다. 이동 전에는 어느 TC 도 이 분기를 안 지났다
