+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/ComposedWidget/DoubleMonthWidget/DoubleMonthWidgetTimlineProvider.swift
@@ -13,6 +13,7 @@ import Optics
 import Domain
 import Extensions
 import CalendarPresentation
+import WidgetScenes
 
 
 // MARK: - widgetViewModel
