+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/ComposedWidget/EventAndMonthWidget/EventAndMonthWidgetTimelineProvider.swift
@@ -13,6 +13,7 @@ import Optics
 import Domain
 import Extensions
 import CalendarPresentation
+import WidgetScenes
 
 
 struct EventAndMonthWidgetViewModel {
