+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/ComposedWidget/TodayAndMonthWidget/TodayAndMonthWidget.swift
@@ -12,6 +12,7 @@ import Domain
 import Extensions
 import CommonPresentation
 import CalendarPresentation
+import WidgetScenes
 
 
 // MARK: - TodayAndMonthWidgetView
