+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/ComposedWidget/TodayAndMonthWidget/TodayAndMonthWidgetTimelineProvider.swift
@@ -13,6 +13,7 @@ import Optics
 import Domain
 import Extensions
 import CalendarPresentation
+import WidgetScenes
 
 
 struct TodayAndMonthWidgetViewModel {
