+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/TodayWidget/TodayWidgetTimelineProvider.swift
@@ -13,6 +13,7 @@ import Optics
 import Domain
 import Extensions
 import CalendarPresentation
+import WidgetScenes
 
 
 struct TodayWidgetTimelineProvider: TimelineProvider {
