+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/TodayWidget/TodayWidgetViewModelProvider.swift
@@ -12,51 +12,13 @@ import Optics
 import Domain
 import Extensions
 import CalendarPresentation
+import WidgetScenes
 
 
-// MARK: - TodayWidgetViewModel
+// MARK: - TodayWidgetViewModel + CalendarEvents
 
-struct TodayWidgetViewModel {
-    
-    let id: CalendarDay
-    let weekDayText: String
-    let day: Int
-    var holidayName: String?
-    var isHoliday: Bool { self.holidayName != nil }
-    let monthAndYearText: String
-    var timeZoneText: String?
-    var todoEventCount: Int = 0
-    var scheduleEventcount: Int = 0
-    var totalEventCount: Int { self.todoEventCount + self.scheduleEventcount }
-    var widgetSetting = WidgetAppearanceSettings()
-    
-    init(
-        id: CalendarDay,
-        weekDayText: String,
-        day: Int,
-        monthAndYearText: String
-    ) {
-        self.id = id
-        self.weekDayText = weekDayText
-        self.day = day
-        self.monthAndYearText = monthAndYearText
-    }
-    
-    init(_ today: Date, _ calendar: Calendar) {
-        self.id = .init(
-            calendar.component(.year, from: today),
-            calendar.component(.month, from: today),
-            calendar.component(.day, from: today)
-        )
-        let timeZone = calendar.timeZone
-        self.weekDayText = today.text("date_form.EEEE".localized(), timeZone: timeZone).uppercased()
-        self.day = calendar.component(.day, from: today)
-        self.monthAndYearText = today.text("date_form.MMM_yyyy".localized(), timeZone: timeZone).uppercased()
-        if timeZone != TimeZone.current {
-            self.timeZoneText = timeZone.localizedName(for: .shortStandard, locale: .current)
-        }
-    }
-    
+extension TodayWidgetViewModel {
+
     func updated(events: CalendarEvents) -> TodayWidgetViewModel {
         let holiday = events.eventWithTimes.compactMap { $0 as? HolidayCalendarEvent }.first
         let todoCount = events.currentTodos.count + events.eventWithTimes.filter { $0 is TodoCalendarEvent }.count
@@ -65,17 +27,6 @@ struct TodayWidgetViewModel {
             |> \.holidayName .~ holiday?.name
             |> \.todoEventCount .~ todoCount
             |> \.scheduleEventcount .~ scheduleCount
-    }
-    
-    static func sample() -> TodayWidgetViewModel {
-        return .init(
-            id: .init(2024, 03, 14),
-            weekDayText: "widget.events.today::sample::sunday".localized(),
-            day: 14,
-            monthAndYearText: "widget.events.today::sample::march2024".localized()
-        )
-        |> \.todoEventCount .~ 3
-        |> \.scheduleEventcount .~ 4
     }
 }
