+++ TodoCalendarApp/AppExtensions/Widget/Tests/ViewModelProviders/MonthWidgetViewModelProviderImpleTests.swift
@@ -14,6 +14,7 @@ import Domain
 import Extensions
 import UnitTestHelpKit
 import TestDoubles
+import WidgetScenes
 
 
 class MonthWidgetViewModelProviderImpleTests: BaseTestCase {
